+++ scripts/ammunition-system/constants.js
@@ -1,7 +1,6 @@
 // Effect IDs
 export const LOADED_EFFECT_ID = "Compendium.pf2e-ranged-combat.effects.nEqdxZMAHlYVXI0Z";
 export const MAGAZINE_LOADED_EFFECT_ID = "Compendium.pf2e-ranged-combat.effects.vKeDaHOu3bGKSk6b";
-export const AMMUNITION_EFFECT_ID = "Compendium.pf2e-ranged-combat.effects.zVcsgX5KduyfBXRZ";
 export const CHAMBER_LOADED_EFFECT_ID = "Compendium.pf2e-ranged-combat.effects.3KT0VeuCOTy5K0lS";
 export const CONJURED_ROUND_EFFECT_ID = "Compendium.pf2e-ranged-combat.effects.psguBnq7jpbnYrhu";
 export const DOUBLE_BARREL_EFFECT_ID = "Compendium.pf2e-ranged-combat.effects.ArViNUKIri5i3E82";